Split test_theory.py into test_theory_general.py and test_theory_helpers.py
Also removes unused phi parameter from pake_pattern signature.
